feat: refactor components to use react-aria-components and remove v2 dependencies
