+++ packages/ui/src/components/Calendar.tsx
@@ -17,10 +17,10 @@ import { useLocale } from "react-aria-components/I18nProvider";
 import { composeRenderProps } from "react-aria-components/composeRenderProps";
 import { cn, tv } from "tailwind-variants";
 
-import { composeTailwindRenderProps } from "../../utils/composeTailwindRenderProps";
-import { focusRing } from "../../utils/focusRing";
-import { Button } from "../Button";
-import { ErrorMessage } from "../form";
+import { Button } from "./Button";
+import { ErrorMessage } from "./form";
+import { composeTailwindRenderProps } from "../utils/composeTailwindRenderProps";
+import { focusRing } from "../utils/focusRing";
 
 const calendarCellVariants = tv({
   extend: focusRing,
