+++ packages/ui/src/components/Select.tsx
@@ -1,141 +1,118 @@
-import { type ComponentPropsWithRef } from "react";
+import { ChevronDown } from "lucide-react";
+import {
+  Select as AriaSelect,
+  type SelectProps as AriaSelectProps,
+  Button,
+  ListBox,
+  type ListBoxProps,
+  type ListBoxItemProps as SelectItemProps,
+  SelectValue,
+  type ValidationResult,
+} from "react-aria-components/Select";
+import { cn, tv } from "tailwind-variants";
 
-import { Check, ChevronDown, ChevronUp } from "lucide-react";
-import { Select as SelectPrimitive, Slot } from "radix-ui";
-import { cn, tv, type VariantProps } from "tailwind-variants";
+import {
+  ListBoxItem as SelectItem,
+  ListBoxSection as SelectSection,
+  type ListBoxSectionProps as SelectSectionProps,
+} from "./ListBox";
+import { Description, FieldError, Label } from "./form";
+import { Popover } from "./v2/Popover";
+import { composeTailwindRenderProps } from "../utils/composeTailwindRenderProps";
+import { focusRing } from "../utils/focusRing";
 
-const {
-  Root: Select,
-  Group: SelectGroup,
-  Value: SelectValue,
-} = SelectPrimitive;
-
-const selectTriggerVariants = tv({
+const selectVariants = tv({
+  extend: focusRing,
   base: [
-    "flex min-h-9 w-full items-center justify-between gap-1 rounded-sm border px-2.5 text-start text-sm/5 transition-colors",
-    "data-placeholder:text-muted-foreground",
-    "data-disabled:cursor-not-allowed data-disabled:opacity-50",
+    "flex min-h-9 w-full min-w-max items-center justify-between gap-1 rounded-sm border border-border bg-surface px-2.5 pr-2 text-start text-sm/5 transition-colors",
+    "[-webkit-tap-highlight-color:transparent]",
   ],
   variants: {
-    variant: {
-      surface: "border bg-surface hover:border-border-hover",
-      outline: "border bg-transparent hover:border-border-hover",
+    isDisabled: {
+      false: [
+        "text-fg-bold group-invalid/select:outline group-invalid/select:outline-destructive hover:border-fg-subtle pressed:border-fg-muted",
+        "forced-colors:group-invalid/select:outline-[Mark]",
+      ],
+      true: [
+        "border-transparent bg-bg-muted text-neutral",
+        "forced-colors:text-[GrayText]",
+      ],
     },
-  },
-  defaultVariants: {
-    variant: "surface",
   },
 });
 
-type SelectTriggerProps = ComponentPropsWithRef<
-  typeof SelectPrimitive.Trigger
-> &
-  VariantProps<typeof selectTriggerVariants>;
+type SelectionMode = "single" | "multiple";
 
-const SelectTrigger = ({
+type SelectProps<T extends object, M extends SelectionMode> = {
+  children: ListBoxProps<T>["children"];
+  label?: string;
+  description?: string;
+  errorMessage?: string | ((validation: ValidationResult) => string);
+  items?: Iterable<T>;
+  listboxProps?: Omit<ListBoxProps<T>, "children">;
+} & Omit<AriaSelectProps<T, M>, "children">;
+
+const Select = <
+  // eslint-disable-next-line @typescript-eslint/no-empty-object-type -- Match `AriaSelectProps` type
+  T extends object = {},
+  M extends SelectionMode = "single",
+>({
   className,
   children,
-  variant,
+  label,
+  description,
+  errorMessage,
+  items,
+  listboxProps,
   ...props
-}: SelectTriggerProps) => (
-  <SelectPrimitive.Trigger
-    className={selectTriggerVariants({ variant, className })}
-    {...props}
-  >
-    <Slot.Slottable>{children}</Slot.Slottable>
-    <SelectPrimitive.Icon asChild>
-      <ChevronDown
-        className="size-4 text-muted-foreground"
-        aria-hidden={true}
-      />
-    </SelectPrimitive.Icon>
-  </SelectPrimitive.Trigger>
-);
-
-const SelectContent = ({
-  className,
-  children,
-  position = "item-aligned",
-  ...props
-}: ComponentPropsWithRef<typeof SelectPrimitive.Content>) => (
-  <SelectPrimitive.Portal>
-    <SelectPrimitive.Content
-      className={cn(
-        "z-[calc(infinity)] overflow-y-auto rounded-sm border bg-surface shadow-xs",
-        className,
-      )}
+}: SelectProps<T, M>) => {
+  return (
+    <AriaSelect
       {...props}
+      className={composeTailwindRenderProps(
+        className,
+        "group/select relative flex flex-col gap-1",
+      )}
     >
-      <SelectPrimitive.ScrollUpButton className="flex h-6.5 cursor-default items-center justify-center">
-        <ChevronUp className="size-4" />
-      </SelectPrimitive.ScrollUpButton>
-      <SelectPrimitive.Viewport
-        className={cn("p-1.5", {
-          "h-(--radix-select-trigger-height) max-h-(--radix-select-content-avaliable-height) w-(--radix-select-trigger-width) max-w-(--radix-select-content-avaliable-width) origin-(--radix-select-content-transform-origin)":
-            position === "popper",
-        })}
-      >
-        <Slot.Slottable>{children}</Slot.Slottable>
-      </SelectPrimitive.Viewport>
-      <SelectPrimitive.ScrollDownButton className="flex h-6.5 cursor-default items-center justify-center">
-        <ChevronDown className="size-4" />
-      </SelectPrimitive.ScrollDownButton>
-    </SelectPrimitive.Content>
-  </SelectPrimitive.Portal>
-);
-
-const SelectItem = ({
-  className,
-  children,
-  ...props
-}: ComponentPropsWithRef<typeof SelectPrimitive.Item>) => (
-  <SelectPrimitive.Item
-    className={cn(
-      "relative flex h-6.5 items-center rounded-sm pr-9.5 pl-6.5 text-sm/none select-none",
-      "data-disabled:cursor-not-allowed data-disabled:opacity-50",
-      "data-highlighted:bg-accent data-highlighted:text-white data-highlighted:outline-none",
-      className,
-    )}
-    {...props}
-  >
-    <SelectPrimitive.ItemIndicator className="absolute left-0 inline-flex w-6.5 items-center justify-center">
-      <Check className="size-4" />
-    </SelectPrimitive.ItemIndicator>
-    <SelectPrimitive.ItemText>
-      <Slot.Slottable>{children}</Slot.Slottable>
-    </SelectPrimitive.ItemText>
-  </SelectPrimitive.Item>
-);
-
-const SelectLabel = ({
-  className,
-  ...props
-}: ComponentPropsWithRef<typeof SelectPrimitive.Label>) => (
-  <SelectPrimitive.Label
-    className={cn("px-6.5 text-xs/6 font-semibold text-foreground", className)}
-    {...props}
-  />
-);
-
-const SelectSeparator = ({
-  className,
-  ...props
-}: ComponentPropsWithRef<typeof SelectPrimitive.Separator>) => (
-  <SelectPrimitive.Separator
-    className={cn("m-1 h-px bg-muted", className)}
-    {...props}
-  />
-);
+      {label && <Label>{label}</Label>}
+      <Button className={selectVariants}>
+        <SelectValue className="basis-content line-clamp-1 shrink-0 text-sm data-[placeholder=true]:text-fg-muted">
+          {({ selectedText, defaultChildren }) =>
+            selectedText || defaultChildren
+          }
+        </SelectValue>
+        <ChevronDown
+          aria-hidden
+          className={cn(
+            "size-4 shrink-0 text-fg-muted group-disabled/select:text-border",
+            "forced-colors:text-[ButtonText] forced-colors:group-disabled/select:text-[GrayText]",
+          )}
+        />
+      </Button>
+      {description && <Description>{description}</Description>}
+      <FieldError>{errorMessage}</FieldError>
+      <Popover containerPadding={0}>
+        <ListBox
+          items={items}
+          className={composeTailwindRenderProps(
+            listboxProps?.className,
+            "max-h-[inherit] overflow-auto p-1 outline-hidden",
+          )}
+          {...listboxProps}
+        >
+          {children}
+        </ListBox>
+      </Popover>
+    </AriaSelect>
+  );
+};
 
 export {
-  selectTriggerVariants,
   Select,
-  SelectTrigger,
-  type SelectTriggerProps,
-  SelectValue,
-  SelectContent,
+  type SelectProps,
+  selectVariants,
   SelectItem,
-  SelectGroup,
-  SelectLabel,
-  SelectSeparator,
+  type SelectItemProps,
+  SelectSection,
+  type SelectSectionProps,
 };
